% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blsa_cardia_arr.R
\docType{data}
\name{blsa_cardia_arr}
\alias{blsa_cardia_arr}
\title{BLSA CARDIA FC Matrices}
\format{An FC matrix array of dimension \eqn{264 \times 264 \times 104}
\describe{
}}
\usage{
blsa_cardia_arr
}
\description{
Array containing the raw FC matrices
}
\keyword{datasets}
