% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blsa_cardia_demo.R
\docType{data}
\name{blsa_cardia_demo}
\alias{blsa_cardia_demo}
\title{BLSA CARDIA Demographics}
\format{A matrix of dimension \eqn{104 \times 7}
\describe{
}}
\usage{
blsa_cardia_demo
}
\description{
Dataset containing the demographics for BLSA and CARDIA subjects
}
\keyword{datasets}
