% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpc_Flury.R
\name{cpc_Flury}
\alias{cpc_Flury}
\title{Common principal components using Flury's (1986) estimation}
\usage{
cpc_Flury(dat, iter = 15)
}
\arguments{
\item{dat}{an array of covariance matrices of dimension \emph{(p x p x N)}}

\item{iter}{numeric, number of iterations}
}
\value{
\code{cpc_Flury} returns a list containing the following components:
\item{D}{eigenvalues for each input matrix as columns}
\item{CPC}{rotation matrix with eigenvectors in columns}
\item{ncomp}{number of components}
}
\description{
Adapted from the FCPCA function in \emph{multigroup} to handle array input
}
