% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpcharmony.R
\name{cpcharmony}
\alias{cpcharmony}
\title{Harmonization using Common Principal Components}
\usage{
cpcharmony(dat, bat)
}
\arguments{
\item{dat}{array of covariance matrices of dimension \eqn{(p \times p \times
N)}}

\item{bat}{batch labels as length \eqn{n} vector}
}
\value{

}
\description{
Harmonization using Common Principal Components
}
