% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcComBat.R
\name{fcComBat}
\alias{fcComBat}
\title{Functional connectivity ComBat}
\usage{
fcComBat(x, bat, eb = TRUE, mod = NULL, to.corr = TRUE, out.pd = FALSE)
}
\arguments{
\item{out.pd}{}
}
\value{

}
\description{
Functional connectivity ComBat
}
