% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcharmony.R
\name{fcharmony}
\alias{fcharmony}
\title{Functional Connectivity Harmonization}
\usage{
fcharmony(
  dat,
  bat,
  mod = NULL,
  method = c("FC-ComBat", "FC-CovBat", "Block-ComBat", "Log-ComBat", "Log-CovBat",
    "PC-ComBat"),
  force.PD = c(FALSE, FALSE),
  to.corr = c(FALSE, FALSE),
  eb = TRUE,
  debug = FALSE
)
}
\arguments{
\item{dat}{An array of covariance matrices of dimension
\emph{(p x p x N)}}

\item{bat}{A vector of batch labels as length \emph{n} vector}

\item{mod}{A matrix of covariates to preserve during harmonization}

\item{method}{Harmonization method. One of "log-CovBat"
or "fc-Combat"}

\item{force.PD}{Logical vector. First element indicates whether input should
be forced to be positive definite using \link[Matrix]{nearPD}, second for
output.}

\item{to.corr}{Logical vector. First element indicates whether input should
be forced to be a correlation matrix using  \link[stats]{cov2cor}, second
for output.}

\item{eb}{Logical, use empirical Bayes in harmonization.}

\item{debug}{Logical, include internal objects in output.}
}
\value{
\code{cpcharmony} returns a list containing the following components:
\item{dat.out}{harmonized data as an array of input dimension}
\item{cpcs}{CPC analysis output list}
\item{cpc.harmony}{harmonization method output}
}
\description{
Functional Connectivity Harmonization
}
