% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_combat.R
\name{log_combat}
\alias{log_combat}
\title{Harmonization using ComBat on log-transformed matrices}
\usage{
log_combat(x, bat, mod = NULL, eb = TRUE, input.pd = FALSE)
}
\arguments{
\item{input.pd}{}
}
\value{

}
\description{
Harmonization using ComBat on log-transformed matrices
}
