% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_covbat.R
\name{log_covbat}
\alias{log_covbat}
\title{Harmonization using CovBat on log-transformed matrices}
\usage{
log_covbat(
  x,
  bat,
  mod = NULL,
  covbat.var = 0.9,
  mean.only = FALSE,
  score.eb = FALSE,
  pc.sym = TRUE,
  input.pd = FALSE
)
}
\arguments{
\item{input.pd}{}
}
\value{

}
\description{
Harmonization using CovBat on log-transformed matrices
}
