% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_combat.R
\name{pc_combat}
\alias{pc_combat}
\title{Harmonization using principal components of matrix}
\usage{
pc_combat(
  x,
  bat,
  mod = NULL,
  covbat.var = 0.9,
  mean.only = FALSE,
  score.eb = FALSE,
  pc.sym = TRUE
)
}
\arguments{
\item{input.pd}{}
}
\value{

}
\description{
Harmonization using principal components of matrix
}
