% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.commTest.R
\name{plot.commTest}
\alias{plot.commTest}
\title{Plot community detections across site}
\usage{
\method{plot}{commTest}(
  x,
  which = 1,
  atlas = NULL,
  atlas.lab = NULL,
  match.labs = FALSE,
  overlap.obj = 1,
  manual.labs = NULL
)
}
\arguments{
\item{x}{\code{commTest} object, typically result of
\link[FCharmony]{test_communities}}

\item{match.labs}{Whether to modify community labels to match across levels.}

\item{overlap.obj}{If matching labels, which objective function to use. All
are sum over all ROIs. 1 is number of overlapping labels, 2 is whether or
not all labels match, 3 is the number of unique labels (minimize), 4 is
a weighted sum between 1 and 2, giving more weight to having all same. 5 is
a weighted sum between 1 and 3, penalizing more distinct labels.}

\item{manual.labs}{Optional vector of length equal to the number of
communities. Recodes community numbers. For example, \code{c(3,2,1)} swaps the
labels of communities 1 and 3.}
}
\value{

}
\description{
Visualize results of community detection output from
\link[FCharmony]{test_communities}
}
