% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_comm_associations.R
\name{test_comm_associations}
\alias{test_comm_associations}
\title{Community detection tests}
\usage{
test_comm_associations(
  x,
  grp,
  comm.method = cluster_louvain,
  metric = "vi",
  threshold = "positive",
  adonis.params = NULL,
  fisher = TRUE,
  debug = FALSE
)
}
\arguments{
\item{x}{\emph{p x p x n} array of functional connectivity matrices or object of
class 'dist', see \link\href{dist}{stats}}

\item{grp}{Factor specifying groups over which to aggregate.}

\item{comm.method}{Function from \code{igraph} that finds communities. See
\link[igraph]{membership} for options. Defaults to
\link[igraph]{cluster_louvain}}

\item{fisher}{Whether to z-transform input FC matrices}

\item{debug}{}

\item{atlas}{Vector of labels specifying original subnetworks.}

\item{grp.method}{Method for obtaining group-level functional connectivity.}

\item{comp.method}{Metric for comparing community detections.}
}
\value{

}
\description{
Examine if grouping affects community detection.
}
