% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_harmony.R
\name{test_harmony}
\alias{test_harmony}
\title{Evaluate harmonization methods}
\usage{
test_harmony(dats, test, force.PD = FALSE, to.corr = FALSE, ...)
}
\arguments{
\item{dats}{Named list of functional connectivity arrays, not necessarily of
the same dimension}

\item{test}{Function to apply to all datasets}

\item{to.corr}{Logical, whether input should
be forced to be a correlation matrix using  \link[stats]{cov2cor}}

\item{...}{Arguments passed to \code{test}}
}
\value{

}
\description{
Apply downstream analysis to multiple datasets, usually outputs of different harmonization methods.
}
