% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mdmr.R
\name{test_mdmr}
\alias{test_mdmr}
\title{Title}
\usage{
test_mdmr(
  ...,
  bat = NULL,
  mod = NULL,
  labs = c("Raw", "Out"),
  tests = c("MDMR", "MDMR-C", "MDMR-L"),
  metric = "E",
  lap.thr = 0.25,
  lap.gam = 0.01,
  mdmr.perm = NULL,
  mdmr.nperm = 10000
)
}
\arguments{
\item{mdmr.perm}{}
}
\value{

}
\description{
Title
}
