% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_norms.R
\name{test_norms}
\alias{test_norms}
\title{Title}
\usage{
test_norms(
  raw,
  out,
  bat,
  tests = c("Original", "Correlation", "Laplacian"),
  metric = "Euclidean",
  lap.thres = 0.5,
  lap.gamma = 0
)
}
\arguments{
\item{bat}{}

\item{metric}{Metric used to calculate distance. Choose from \code{method} in
\link[shapes]{distcov}}
}
\value{

}
\description{
Title
}
