% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_regress.R
\name{test_regress}
\alias{test_regress}
\title{Title}
\usage{
test_regress(
  ...,
  bat = NULL,
  mod = NULL,
  roi.names = NULL,
  labs = c("Raw", "Out"),
  tests = c("Elem", "Group", "CPC"),
  cpc.k = 15,
  fisher = TRUE,
  seed.roi = NULL,
  to.corr = FALSE,
  debug = FALSE
)
}
\arguments{
\item{to.corr}{}
}
\value{

}
\description{
Title
}
