% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blComBat.R
\name{blComBat}
\alias{blComBat}
\title{Experimental method for harmonization while accounting for subnetworks}
\usage{
blComBat(
  x,
  bat,
  mod = NULL,
  blocks = dimnames(x)[[1]],
  eb = TRUE,
  to.corr = TRUE,
  out.pd = FALSE,
  fisher = TRUE
)
}
\arguments{
\item{blocks}{Blocks indicating submatrices that are harmonized in the
second-stage ComBat. Defaults to input dimension names.}

\item{out.pd}{}
}
\value{

}
\description{
Global ComBat followed by within-block ComBat
}
