% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpcharmony.R
\name{cpcharmony}
\alias{cpcharmony}
\title{Harmonization using Common Principal Components}
\usage{
cpcharmony(
  dat,
  bat,
  mod = NULL,
  log.input = FALSE,
  log.dat = NULL,
  cpc.method = c("stepwise", "Flury", "CAP", "none"),
  cpc.k = dim(dat)[1],
  err.method = c("log-CovBat", "log-ComBat", "ComBat", "CovBat", "PC-ComBat",
    "log-PC-ComBat", "remove", "none"),
  method = c("ComBat", "log-ComBat", "None"),
  force.PD = c(FALSE, FALSE),
  to.corr = c(FALSE, FALSE),
  cpc.cap.oc = FALSE,
  debug = FALSE
)
}
\arguments{
\item{dat}{An array of covariance matrices of dimension
\emph{(p x p x N)}}

\item{bat}{A vector of batch labels as length \emph{n} vector}

\item{mod}{A matrix of covariates to preserve during harmonization}

\item{cpc.method}{Estimation method for common principal components. One of
"stepwise", "Flury", or "CAP". Otherwise, skips step.}

\item{cpc.k}{Numeric, number of common principal components to harmonize}

\item{err.method}{Harmonization method for error terms. One of "log-CovBat"
or "fc-Combat"}

\item{method}{Method for harmonization. One of "likelihood", "log-ComBat",
"ComBat". Currently only ComBat supported.}

\item{force.PD}{Logical vector. First element indicates whether input should
be forced to be positive definite using \link[Matrix]{nearPD}, second for
output}

\item{to.corr}{Logical vector. First element indicates whether input should
be forced to be a correlation matrix using  \link[stats]{cov2cor}, second
for output}

\item{cpc.cap.oc}{Logical, only used if \code{cpc.method = "CAP"}, if
\code{TRUE} imposes orthogonal constraint when identifiying higher-order PCs.}
}
\value{
\code{cpcharmony} returns a list containing the following components:
\item{dat.out}{harmonized data as an array of input dimension}
\item{cpcs}{CPC analysis output list}
\item{cpc.harmony}{harmonization method output}
}
\description{
Harmonization using Common Principal Components
}
