% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcCovBat.R
\name{fcCovBat}
\alias{fcCovBat}
\title{Functional connectivity CovBat}
\usage{
fcCovBat(
  x,
  bat,
  eb = TRUE,
  mod = NULL,
  to.corr = TRUE,
  out.pd = FALSE,
  fisher = TRUE
)
}
\arguments{
\item{out.pd}{}
}
\value{

}
\description{
Functional connectivity CovBat
}
