% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_communities.R
\name{test_communities}
\alias{test_communities}
\title{Community detection tests
Examine if site effects exist in community detection}
\usage{
test_communities(
  ...,
  labs = c("Raw", "Out"),
  bat,
  atlas,
  grp.method = c("average", "multilayer"),
  comm.method = cluster_louvain,
  metric = c("adjusted.rand", "nmi"),
  threshold = "positive",
  fisher = TRUE,
  debug = FALSE
)
}
\arguments{
\item{atlas}{Vector of labels specifying original subnetworks.}

\item{grp.method}{Method for obtaining group-level functional connectivity.}

\item{comm.method}{Function from \code{igraph} that finds communities. See
\link[igraph]{membership} for options. Defaults to
\link[igraph]{cluster_louvain}}

\item{fisher}{Whether to z-transform input FC matrices}

\item{debug}{}

\item{comp.method}{Metric for comparing community detections.}
}
\value{

}
\description{
Community detection tests
Examine if site effects exist in community detection
}
