% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_net.R
\name{test_net}
\alias{test_net}
\title{Functional Connectivity Evaluations
Many proposed by Meichen Yu. Refer to the 2018 paper for further details:
\url{https://doi.org/10.1002/hbm.24241}}
\usage{
test_net(
  ...,
  bat = NULL,
  net.rois = NULL,
  net.cov = NULL,
  labs = c("Raw", "Out"),
  net.method = c("corr", "pcorr", "glasso"),
  net.metrics = c("global", "local", "within", "modularity", "part.coeff"),
  kw.p = FALSE,
  threshold = NULL,
  fisher = TRUE,
  mod.clust = cluster_louvain,
  weighted = TRUE,
  glasso.rho = NULL,
  debug = FALSE
)
}
\arguments{
\item{threshold}{Function that takes adjacency matrix input and returns an
adjacency matrix with thresholded values. Examples include `x > 0.5` for
right-tail thresholding and `x[x <= 0.5] = 0; x` for right-tail thresholding
while keeping weights}

\item{mod.clust}{Function from `igraph` that finds communities. See
\link[igraph]{membership} for options. Defaults to
\link[igraph]{cluster_louvain}, which is used in the Brain Connectivity
Toolbox}

\item{out.graphs}{}
}
\value{

}
\description{
Functional Connectivity Evaluations
Many proposed by Meichen Yu. Refer to the 2018 paper for further details:
\url{https://doi.org/10.1002/hbm.24241}
}
