% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_norms.R
\name{test_norms}
\alias{test_norms}
\title{Title}
\usage{
test_norms(
  raw,
  out,
  ...,
  bat = NULL,
  labs = c("Raw", "Out"),
  tests = c("Original", "Correlation", "Laplacian", "MDMR"),
  metric = "E",
  lap.thr = 0.25,
  lap.gam = 0.01,
  mdmr.perm = NULL
)
}
\arguments{
\item{bat}{}

\item{metric}{Metric used to calculate distance. Choose from \code{method} in
\link[CovTools]{CovDist}.}
}
\value{

}
\description{
Title
}
