% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_regress.R
\name{test_regress}
\alias{test_regress}
\title{Title}
\usage{
test_regress(
  raw,
  out,
  ...,
  bat = NULL,
  mod = NULL,
  labs = c("Raw", "Out"),
  tests = c("MDMR", "MDMR-C", "MDMR-L"),
  metric = "E",
  lap.thr = 0.25,
  lap.gam = 0.01,
  mdmr.perm = NULL
)
}
\arguments{
\item{mdmr.perm}{}
}
\value{

}
\description{
Title
}
