% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_yu.R
\name{test_yu}
\alias{test_yu}
\title{Functional Connectivity Evaluations proposed by Meichen Yu
Refer to the 2018 paper for further details:
\url{https://doi.org/10.1002/hbm.24241}}
\usage{
test_yu(
  ...,
  bat = NULL,
  net.rois = NULL,
  net.cov = NULL,
  threshold = NULL,
  labs = c("Raw", "Out"),
  p.method = "BH",
  net.method = c("corr", "pcorr", "glasso"),
  net.weighted = TRUE,
  net.only = TRUE,
  glasso.rho = NULL
)
}
\arguments{
\item{bat}{}
}
\value{

}
\description{
Functional Connectivity Evaluations proposed by Meichen Yu
Refer to the 2018 paper for further details:
\url{https://doi.org/10.1002/hbm.24241}
}
